refactor: fix create-only semantics and dedupe private parseSchema
Code review follow-ups:

- POST /company-verification/companies now returns 409 when a caller
  supplies an id that already exists, instead of silently upserting
  (Company ids are referenced by user_company.companyId, so overwriting
  an existing company is destructive). The conflict is checked before
  the Cloudinary upload to avoid wasted work.
- Extract the duplicated parseSchema helper into src/routes/private/utils.ts
  and import it from both contributions and companyVerification routers
  (CLAUDE.md single-source-of-truth guidance).
